+++ pos_scan_ref/static/src/js/pos.js
@@ -1,52 +1,76 @@
-(function(){
+/*  Copyright 2014 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+    Copyright 2019 Artem Rafailov <https://it-projects.info/team/Ommo73>
+    License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html). */
+odoo.define('pos_scan_ref.pos', function (require) {
+"use strict";
 
-function pos(instance, module){
-    module.BarcodeReader.include({
-        scan: function(code){
-            if(code.length < 3){
-                return;
-            }else if(this.pos.db.get_product_by_reference(code)){
-                var parse_result = {
-                    encoding: 'reference',
-                    type: 'product',
-                    code: code,
-                };
-            }else if(code.length === 13 && this.check_ean(code)){
-                var parse_result = this.parse_ean(code);
-            }else if(code.length === 12 && this.check_ean('0'+code)){
-                // many barcode scanners strip the leading zero of ean13 barcodes.
-                // This is because ean-13 are UCP-A with an additional zero at the beginning,
-                // so by stripping zeros you get retrocompatibility with UCP-A systems.
-                var parse_result = this.parse_ean('0'+code);
-            }else{
-                var parse_result = {
-                    encoding: 'error',
-                    type: 'error',
-                    code: code,
-                };
+var devices = require('point_of_sale.devices');
+var PosDb = require('point_of_sale.DB');
+var models = require('point_of_sale.models');
+
+var PosModelSuper = models.PosModel;
+models.PosModel = models.PosModel.extend({
+    scan_product: function(parsed_code){
+        var selectedOrder = this.get_order();
+        var product = this.db.get_product_by_reference(parsed_code.code);
+        if(parsed_code.encoding === 'reference'){
+            if(!product){
+                return false;
             }
-
-            if(parse_result.type in {'product':'', 'weight':'', 'price':'', 'discount':''}){    //ean is associated to a product
-                if(this.action_callback.product){
-                    this.action_callback.product(parse_result);
-                }
+            if(parsed_code.type === 'price'){
+                selectedOrder.add_product(product, {price:parsed_code.value});
+            }else if(parsed_code.type === 'weight'){
+                selectedOrder.add_product(product, {quantity:parsed_code.value, merge:false});
+            }else if(parsed_code.type === 'discount'){
+                selectedOrder.add_product(product, {discount:parsed_code.value, merge:false});
             }else{
-                if(this.action_callback[parse_result.type]){
-                    this.action_callback[parse_result.type](parse_result);
-                }
+                selectedOrder.add_product(product);
+            }
+            return true;
+        }
+        return PosModelSuper.prototype.scan_product.apply(this, arguments);
+    },
+});
+
+devices.BarcodeReader.include({
+    scan: function(code){
+        var parsed_result = {};
+        if(this.pos.db.get_product_by_reference(code)){
+            parsed_result = {
+                encoding: 'reference',
+                type: 'product',
+                code: code,
+            };
+            this.action_callback[parsed_result.type](parsed_result);
+        }else if(this.pos.db.get_product_by_reference('0'+code)){
+            parsed_result = {
+                encoding: 'reference',
+                type: 'product',
+                code: '0'+code,
+            };
+            this.action_callback[parsed_result.type](parsed_result);
+        }else{
+            this._super(code);
+        }
+    }
+});
+
+PosDb.include({
+    init: function(options){
+        this._super(options);
+        this.product_by_reference = {};
+    },
+    add_products: function(products){
+        this._super(products);
+        for(var i = 0, len = products.length; i < len; i++){
+            var product = products[i];
+            if(product.default_code){
+                this.product_by_reference[product.default_code] = product;
             }
         }
-    });
-}
-
-var _super = window.openerp.point_of_sale;
-window.openerp.point_of_sale = function(instance){
-    _super(instance);
-    var module = instance.point_of_sale;
-
-    pos(instance, module);
-
-    //$('<link rel="stylesheet" href="/pos_product_lot/static/src/css/pos.css"/>').appendTo($("head"))
-};
-
-})();
+    },
+    get_product_by_reference: function(ref){
+        return this.product_by_reference[ref];
+    },
+});
+});